Refactored some code in factored frontier that runs the inner BP loop into its own trait so it can be shared between other algorithms that use BP as an inner loop (e.g., particle BP).
